requirements.txt
fastapi==0.98.0
pydantic==1.10.9
uvicorn==0.22.0
uvloop==0.17.0
orjson==3.9.1
python-dotenv==1.0.0

sqlalchemy==2.0.17
psycopg2-binary==2.9.6
asyncpg==0.27.0

pytest==7.4.0
httpx==0.24.1
pytest-asyncio==0.21.0

python-jose==3.3.0

alembic==1.11.1

python-multipart==0.0.6
redis==4.6.0